Перенес модули в директорию /modules
