+++ backend/src/modules/auth/interceptors/photo-interceptor.ts
