+++ backend/src/modules/auth/strategies/jwt.strategy.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { ExtractJwt, Strategy } from 'passport-jwt';
 import { UserDto } from '#server/common/dto/user.dto';
-import { UsersService } from '#src/user/users.service';
+import { UsersService } from '#src/modules/user/users.service';
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy) {
