+++ backend/src/modules/items/items.module.ts
@@ -2,9 +2,9 @@ import { Module } from '@nestjs/common';
 import { ItemsService } from './items.service';
 import { ItemsController } from './items.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { ItemEntity } from '#src/items/entity/item.entity';
-import { PhotoEntity } from '#src/photos/entity/photo.entity';
-import { CategoryEntity } from '#src/items/entity/category.entity';
+import { ItemEntity } from '#src/modules/items/entity/item.entity';
+import { PhotoEntity } from '#src/modules/photos/entity/photo.entity';
+import { CategoryEntity } from '#src/modules/items/entity/category.entity';
 
 @Module({
   imports: [
