+++ backend/src/modules/uploads/dto/get-image-params.dto.ts
